# Use memchr for finding end of line