--- a/Cargo.toml
+++ b/Cargo.toml
@@ -15,3 +15,4 @@ categories = ["asynchronous", "network-programming"]
 [dependencies]
 bytes = "0.4.12"
 futures-preview = "0.3.0-alpha.17"
+memchr = "2.2.1"