Benchmarking results on my MacBook Air with 1,6 GHz Intel Core i5 (one of three runs each, runs were within an eyeballed 1% of each other):

Before:

```
test long   ... bench:      10,252 ns/iter (+/- 401)
test medium ... bench:       1,742 ns/iter (+/- 78)
test short  ... bench:       1,082 ns/iter (+/- 94)
```

After:

```
test long   ... bench:       1,895 ns/iter (+/- 267)
test medium ... bench:       1,170 ns/iter (+/- 72)
test short  ... bench:         997 ns/iter (+/- 90)
```